проходит пфт_2019, кроме УБ, проходит юнит тесты, все круто и классно, надо приводить в норму

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: grvelva <grvelva@student.21-school.ru>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/22 18:28:40 by grvelva           #+#    #+#             */
-/*   Updated: 2020/12/13 18:03:04 by grvelva          ###   ########.fr       */
+/*   Updated: 2020/12/16 13:47:16 by grvelva          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ void		putchar_right(size_t len, char c, char empty);
 size_t		putchar_len(s_output *frmt);
 int			print_char(char c, s_output *frmt);
 int			print_ptr(va_list args, s_output *frmt);
-char		*ft_ptrtostr(long long int nbr, char *base);
+char		*ft_ptrtostr(unsigned long nbr, char *base);
 int			print_nbr(va_list args, s_output *frmt);
 int			print_unbr(va_list args, s_output *frmt);
 int			ft_putnbr(long long i);
