+++ print_unbr.c
@@ -18,7 +18,7 @@ int		print_unbr(va_list args, s_output *frmt)
 	char			empty;
 	int 			len;
 
-	nbr = va_arg(args, long long);
+	nbr = va_arg(args, unsigned int);
 	(frmt->precision > -1) ? frmt->flag = frmt->flag & 15 : frmt->flag;
 	empty = ((frmt->flag & ZERO) && frmt->precision == -1) ? '0' : ' ';
 	if (frmt->precision == 0 && nbr == 0)
